+++ changelog.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## Upcoming
+* Added Red block outline with wrong tool
+  * If the tool in hand will make the block drop no items then the block outline will be green tinted
 * Fixed mobs arrows being affected by arrows damage multiplier
 
 ## 3.13.2